requirements.txt
django==3.2
djangorestframework==3.13.1
matplotlib==3.5.2